Create lesson form and DRY lesson backoffice
