+++ .github/workflows/test.yml
@@ -36,8 +36,8 @@ jobs:
           - 27017:27017
         options: >-
           --health-cmd "mongosh --eval 'db.adminCommand(\"ping\")'"
-          --health-interval 10s
-          --health-timeout 5s
+          --health-interval 1s
+          --health-timeout 1s
           --health-retries 5
 
     steps:
